requirements.txt
python-telegram-bot==20.7
Flask==3.1.1
gspread==6.2.1
oauth2client==4.1.3
